Agregadas las modificaciones y creado el primer test, está para revisar pero suficiente closes #9
