+++ test_zip_utilities.py
@@ -1,13 +1,17 @@
 # test_mi_modulo.py
 import unittest
  
-from ExtraerJSONS import listar_archivos_zip
-from ExtraerJSONS import crear_ficheros_json_del_zip
+from zip_utils import listar_archivos_zip,crear_ficheros_json_del_zip
 
 class TestListarArchivosZip(unittest.TestCase):
 
     def test_listar_archivos_zip(self):
-        self.fail("Not implemented")
+
+        expected = ["test_helper/test_zip_utilities\\file1.zip","test_helper/test_zip_utilities\\file4.ZIP"]
+        target = listar_archivos_zip("test_helper/test_zip_utilities")
+
+
+        self.assertEquals(expected,target)
     
     def test_crear_ficheros_json_del_zip(self):
         self.fail("Not implemented")
